combine `env` and `runMode` (#368)
* combine env and runMode

* remove async from functions that no longer use await

* Suggestion to decouple the event callback setup from the database setup.

We're able to capture the database being ready in a promise variable, which we can then `await` on the specific node callbacks that actually care about the database/logic files (which is admittedly _most of them_, but being able to do this decoupling for the ClientReady is a bit of a performance boost in being able to go through the motions of getting logged in and posting basic update information while the database is doing some transactions. We also avoid a pitfall of implying that the final `catch` is able to handle the errors produced within the callback functions set on the client events. That `catch` would be able to handle any errors that happen as those callbacks are set on those events, but once those callbacks are set, the error handling is within a different calling context (but not variable/block context) within the Node session.

* include improvements from combineEnvAndRunMode2
- use || to default runMode to "development" in case of empty string
- fix global scope iterator in setDB iteration
- rename `sequelize` to `dbConnection`
- rename `client` to `dAPIClient`

---------

Co-authored-by: SillySalamndr <l.ensign@hotmail.com>